pyproject.toml
[project]
name = "foundry-tui"
version = "0.1.0"
description = "A terminal UI for testing Azure AI Foundry models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.47.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyperclip>=1.8.0",
]

[project.scripts]
foundry-tui = "foundry_tui.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/foundry_tui"]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that hit real Azure Foundry endpoints (need credentials in .env)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
